+++ cdoc2-cli/src/main/java/ee/cyber/cdoc2/cli/commands/CDocDecryptCmd.java
@@ -70,7 +70,8 @@ public class CDocDecryptCmd implements Callable<Void> {
     }
 
     @Option(names = {"-c", "--crypto-stick"},
-        description = "Specify what type of crypto stick is used, allowed values: [SECP256R1, SECP384R1]")
+        description = "Specify what type of crypto stick is used, allowed values: "
+            + "[SECP256R1, SECP384R1, RSA3072, RSA4096]")
     private CryptoStickConf cryptoStickConf;
 
     @Override