+++ cdoc2-lib/src/main/java/ee/cyber/cdoc2/CryptoStickConf.java
@@ -5,7 +5,9 @@ package ee.cyber.cdoc2;
  */
 public enum CryptoStickConf {
     SECP256R1(256),
-    SECP384R1(384);
+    SECP384R1(384),
+    RSA3072(3072),
+    RSA4096(4096);
 
     private final int keySize;